avoid relying on undefined behavoiur, fails with clang

+++ src/t_string.c
@@ -40,7 +40,8 @@ int getGenericCommand(client *c);
 static int checkStringLength(client *c, long long size, long long append) {
     if (mustObeyClient(c))
         return C_OK;
-    long long total = size + append;
+    /* 'uint64_t' cast is there just to prevent undefined behavior on overflow */
+    long long total = (uint64_t)size + append;
     /* Test configured max-bulk-len represending a limit of the biggest string object,
      * and also test for overflow. */
     if (total > server.proto_max_bulk_len || total < size || total < append) {
